+++ test-api/schema-paginated/User.js
@@ -167,11 +167,11 @@ const User = new GraphQLObjectType({
           ...do {
             if (PAGINATE === 'offset') {
               ;({
-                orderBy: args => ({
+                orderBy: args => [
                   // eslint-disable-line no-unused-vars
-                  created_at: 'desc',
-                  id: 'asc'
-                })
+                  { column: 'created_at', direction: 'desc' },
+                  { column: 'id', direction: 'asc' }
+                ]
               })
             } else if (PAGINATE === 'keyset') {
               ;({
@@ -235,10 +235,10 @@ const User = new GraphQLObjectType({
               ;({
                 orderBy: args =>
                   args.sortOnMain
-                    ? {
-                        created_at: 'ASC',
-                        id: 'ASC'
-                      }
+                    ? [
+                        { column: 'created_at', direction: 'ASC' },
+                        { column: 'id', direction: 'ASC' }
+                      ]
                     : null
               })
             } else if (PAGINATE === 'keyset') {
@@ -391,10 +391,10 @@ const User = new GraphQLObjectType({
           ...do {
             if (PAGINATE === 'offset') {
               ;({
-                orderBy: {
-                  id: 'ASC',
-                  created_at: 'ASC'
-                }
+                orderBy: [
+                  { column: 'id', direction: 'ASC' },
+                  { column: 'created_at', direction: 'ASC' }
+                ]
               })
             } else if (PAGINATE === 'keyset') {
               ;({