+++ test-d/index.test-d.ts
@@ -31,10 +31,10 @@ new GraphQLObjectType<any, ExampleContext>({
           ignoreAll: true,
           ignoreTable: true,
           limit: 10,
-          orderBy: {
-            foo: 'ASC',
-            bar: 'DESC'
-          },
+          orderBy: [
+            { column: 'foo', direction: 'asc' },
+            { column: 'bar', direction: 'DESC' }
+          ],
           sortKey: {
             order: 'ASC',
             key: ['id']